+++ ClientProject/utility/RingBuffer.cpp
@@ -179,7 +179,7 @@ namespace utility
     void RingBuffer::MoveRear(ringBufferSize iSize)
     {
         char *pChk;
-        long long distance;
+        int64_t distance;
         char *oldRear;
 
         oldRear = mRearPtr;
@@ -195,7 +195,7 @@ namespace utility
     void RingBuffer::MoveFront(ringBufferSize iSize)
     {
         char *pChk;
-        long long distance;
+        int64_t distance;
         char *oldFront;
 
         oldFront = mFrontPtr;
