+++ _lib/AcceptEx_IOCP_NetworkLib/utility/RingBuffer.cpp
@@ -191,7 +191,7 @@ namespace utility
     void RingBuffer::MoveRear(ringBufferSize iSize)
     {
         char *pChk;
-        long long distance;
+        int64_t distance;
         char *oldRear;
 
         oldRear = mRearPtr;
@@ -207,7 +207,7 @@ namespace utility
     void RingBuffer::MoveFront(ringBufferSize iSize)
     {
         char *pChk;
-        long long distance;
+        int64_t distance;
         char *oldFront;
 
         oldFront = mFrontPtr;
